Strip index.html from links to tags
Implement a custom tagsField context (copied from the Hakyll source) which
strips the file name from the URL in links to tags.

This is kind of a pain in the arse; I rather expected Hakyll to provide a
better way to do this. Maybe it's time for a think and a patch?