Use Hakyll's pagination facilities for archive pages
This generates the paginated archives, but there are too many problems for this
to be a useful solution. In particular, there's no way to control the ordering
of items.

I really need to find a generic way to deal with the trailing index.html in
URLs without having to handle each case (posts, pages and now pagination links)
with its own hack.